# apps/projects-service/openapi.json
{"openapi":"3.0.0","paths":{"/api/projects":{"get":{"operationId":"ProjectsController_isServerUp","parameters":[{"name":"Projects-API","in":"header","description":"Projects related endpoints","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/isServerUpDTO"}}}}},"tags":["projects"]},"post":{"operationId":"ProjectsController_createProject","parameters":[{"name":"Projects-API","in":"header","description":"Projects related endpoints","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectRequestDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponseDTO"}}}}},"tags":["projects"]}},"/api/projects/{id}":{"get":{"operationId":"ProjectsController_getAllUserProjects","parameters":[{"name":"Projects-API","in":"header","description":"Projects related endpoints","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","description":"string for the user id","schema":{"oneOf":[{"type":"string"},{"type":"integer"}]}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllUserProjectsDTO"}}}}},"tags":["projects"]},"put":{"operationId":"ProjectsController_updateProject","parameters":[{"name":"Projects-API","in":"header","description":"Projects related endpoints","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","description":"string for the project id","schema":{"oneOf":[{"type":"string"},{"type":"integer"}]}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectRequestDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponseDTO"}}}}},"tags":["projects"]},"delete":{"operationId":"ProjectsController_deleteProject","parameters":[{"name":"Projects-API","in":"header","description":"Projects related endpoints","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","description":"string for the project id","schema":{"oneOf":[{"type":"string"},{"type":"integer"}]}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteProjectResDTO"}}}}},"tags":["projects"]}},"/api/projects/project/{projectId}":{"get":{"operationId":"ProjectsController_getUserProjectById","parameters":[{"name":"Projects-API","in":"header","description":"Projects related endpoints","schema":{"type":"string"}},{"name":"projectId","required":true,"in":"path","description":"string for the project id","schema":{"oneOf":[{"type":"string"},{"type":"integer"}]}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponseDTO"}}}}},"tags":["projects"]}}},"info":{"title":"Projects Service","description":"Projects Service api documentation","version":"1.0","contact":{}},"tags":[{"name":"projects-service","description":""}],"servers":[],"components":{"schemas":{"isServerUpDTO":{"type":"object","properties":{"message":{"type":"string","description":"message"}},"required":["message"]},"CreateProjectRequestDTO":{"type":"object","properties":{"userId":{"type":"string","description":"User id"},"genre":{"type":"string","description":"Project genre"}},"required":["userId","genre"]},"UserInfo":{"type":"object","properties":{"userId":{"type":"string","description":"User ID"},"username":{"type":"string","description":"Username"},"img":{"type":"string","description":"User Image"}},"required":["userId","username","img"]},"ProjectResponseDTO":{"type":"object","properties":{"_id":{"type":"string","description":"Project id"},"userInfo":{"description":"User info","allOf":[{"$ref":"#/components/schemas/UserInfo"}]},"name":{"type":"string","description":"Project name","minimum":2,"maximum":40},"img":{"type":"string","description":"Project img","default":""},"genre":{"type":"string","description":"Project genre"},"description":{"type":"string","description":"Project description","maximum":300,"default":""},"shared":{"type":"boolean","description":"Project share status","default":true}},"required":["_id","userInfo","name","img","genre","description","shared"]},"GetAllUserProjectsDTO":{"type":"object","properties":{"projects":{"description":"Projects","type":"array","items":{"$ref":"#/components/schemas/ProjectResponseDTO"}}},"required":["projects"]},"UpdateProjectRequestDTO":{"type":"object","properties":{"name":{"type":"string","description":"Project name","minimum":2,"maximum":40},"img":{"type":"string","description":"Project img","default":""},"description":{"type":"string","description":"Project description","maximum":300,"default":""},"shared":{"type":"boolean","description":"Project share status","default":true}}},"DeleteProjectResDTO":{"type":"object","properties":{"_id":{"type":"string","description":"project id"}},"required":["_id"]}}}}